page/dl_recommendation/requirements.txt
keras
tensorflow
matplotlib
requests
eyed3
Pillow
subprocess.run
sklearn
